change app driectory strcuture and add sample files and code
